+++ R/plotPA.R
@@ -74,7 +74,7 @@ plotPA <- function(map, th, colors = NULL, hr = FALSE, filename = NULL,
     theme(plot.title = element_text(hjust = 0.5),
           axis.ticks.x = element_blank(),
           axis.ticks.y = element_blank(),
-          text = element_text(colour = "#666666", family = "sans-serif"))
+          text = element_text(colour = "#666666"))
 
   return(my_plot)
 }