+++ R/plotPred.R
@@ -56,7 +56,7 @@ plotPred <- function(map, lt = "", colorramp = NULL, hr = FALSE) {
     theme(plot.title = element_text(hjust = 0.5),
           axis.ticks.x = element_blank(),
           axis.ticks.y = element_blank(),
-          text = element_text(colour = "#666666", family = "sans-serif"))
+          text = element_text(colour = "#666666"))
 
   return(my_plot)
 }