+++ NEWS.md
@@ -1,6 +1,7 @@
 # SDMtune 0.2.0.9000
 * New function to predict SDMmodelCV objects
 * Add the possibility to pass multiple methods to the `train` function
+* Add possibility to compute the testing AUC or TSS for a held apart testing dataset in the case of a `SDMmodelCV` object
 * Add the possibility to merge only the presence locations in the `mergeSWD` function
 
 # SDMtune 0.2.0